(Windows) Added the libgit2 library as a dependency

+++ tools/Windows_dependencies_setup_MinGW.sh
@@ -71,6 +71,21 @@ fi
 
 cp -p FreeImage/Dist/x64/FreeImage.dll ..
 
+echo -e "\nSetting up libgit2"
+rm -rf libgit2
+
+git clone https://github.com/libgit2/libgit2.git
+
+if [ ! -d libgit2 ]; then
+  echo "libgit2 directory is missing, aborting."
+  exit
+fi
+
+cd libgit2
+git checkout v1.6.3
+mkdir build
+cd ..
+
 echo -e "\nSetting up pugixml"
 rm -rf pugixml
